\name{pearson}
\docType{data}
\alias{pearson}
\title{Karl Pearson's lifetime published articles}
\description{
 List of Karl Pearson's article publications over the course of his career. Includes all subjects from journals and dates covered by JSTOR database \url{http://www.jstor.org}
}
\usage{pearson}
\format{
  Reference list output from \code{\link{read.ris}}
}

\source{
 JSTOR
}

\references{
Porter, T. M. (2004). \emph{Karl Pearson: The scientific life in a statistical age.} Princeton: Princeton University Press.
}

\keyword{datasets}
