\name{ris-package}
\alias{ris-package}
\alias{ris}
\docType{package}
\title{Package for bibliographic analysis}
\description{
Functions to import and analyze references from an RIS-formatted source
}
\details{
\tabular{ll}{
Package: \tab ris\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-01-24\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
Import RIS files with read.ris, convert list of references to citation objects
}

\author{
Author/Maintainer: S. Kovalchik \email{skoval@ucla.edu}
}
\references{}
\keyword{file}
\keyword{connection}

\keyword{ package }
\seealso{}
